Move mini-macro to auxilary
